on ARP lookup failures when forwarding packets, just drop the packet
I.writev will throw the exception Ipv4.Routing.No_route_to_destination_address when
ARP lookup of a destination address fails.  The exception (uncatchable
in the current release of tcpip, since Ipv4.Routing is not exposed in
ipv4.mli) halts the program.  This is inappropriate for a NAT device,
which should be able to survive clients on the network dropping their
links!  Catch this exception (relying on a patch to mirage-tcpip to
expose it) and respond by dropping the packet.